separating out types to resolve cyclic dependencies on EE
